+++ src/components/legend/legend-list/legend-item/legend-item-buttons/legend-item-button-info/index.js
@@ -1,3 +1,46 @@
-import LegendItemButtonInfo from './legend-item-button-info-component';
+import React, { PureComponent } from 'react';
+import PropTypes from 'prop-types';
+import CSSModules from 'react-css-modules';
 
-export default LegendItemButtonInfo;
+// Components
+import Icon from 'components/icon';
+
+// Tooltip
+import Tooltip from 'rc-tooltip/dist/rc-tooltip';
+
+// Styles
+import styles from '../styles-button.scss';
+
+class LegendItemButtonInfo extends PureComponent {
+  static propTypes = {
+    activeLayer: PropTypes.object.isRequired,
+    onChangeInfo: PropTypes.func.isRequired
+  }
+
+  render() {
+    const { activeLayer } = this.props;
+
+    return (
+      <Tooltip
+        overlay="Info"
+        overlayClassName="c-rc-tooltip -default"
+        overlayStyle={{ color: '#fff' }}
+        placement="top"
+        trigger={['hover', 'click']}
+        mouseLeaveDelay={0}
+        destroyTooltipOnHide
+      >
+        <button
+          type="button"
+          styleName="c-legend-button"
+          aria-label="More information"
+          onClick={() => this.props.onChangeInfo(activeLayer)}
+        >
+          <Icon name="icon-info" className="-small" />
+        </button>
+      </Tooltip>
+    );
+  }
+}
+
+export default CSSModules(LegendItemButtonInfo, styles, { allowMultiple: true });
