--- src/components/legend/legend-list/legend-item/legend-item-buttons/legend-item-button-remove/legend-item-button-remove-component.js
@@ -1,46 +1,0 @@
-import React, { PureComponent } from 'react';
-import PropTypes from 'prop-types';
-
-// Components
-import Icon from 'components/icon';
-
-// Tooltip
-import Tooltip from 'rc-tooltip/dist/rc-tooltip';
-
-class LegendItemButtonRemove extends PureComponent {
-  static propTypes = {
-    activeLayer: PropTypes.object,
-
-    // ACTIONS
-    onRemoveLayer: PropTypes.func
-  }
-
-  render() {
-    const { activeLayer } = this.props;
-
-    return (
-      <Tooltip
-        overlay="Remove"
-        overlayClassName="c-rc-tooltip -default"
-        overlayStyle={{
-          color: '#fff'
-        }}
-        placement="top"
-        trigger={['hover', 'click']}
-        mouseLeaveDelay={0}
-        destroyTooltipOnHide
-      >
-        <button
-          type="button"
-          className="close"
-          onClick={() => this.props.onRemoveLayer(activeLayer)}
-          aria-label="Remove"
-        >
-          <Icon name="icon-cross" />
-        </button>
-      </Tooltip>
-    );
-  }
-}
-
-export default LegendItemButtonRemove;
