--- src/components/legend/legend-list/legend-item/legend-item-buttons/legend-item-button-visibility/legend-item-button-visibility-component.js
@@ -1,45 +1,0 @@
-import React, { PureComponent } from 'react';
-import PropTypes from 'prop-types';
-
-// Components
-import Icon from 'components/icon';
-
-// Tooltip
-import Tooltip from 'rc-tooltip/dist/rc-tooltip';
-
-class LegendItemButtonVisibility extends PureComponent {
-  static propTypes = {
-    activeLayer: PropTypes.object,
-    visible: PropTypes.bool,
-    onChangeVisibility: PropTypes.func
-  }
-
-  render() {
-    const { activeLayer, visible } = this.props;
-
-    return (
-      <Tooltip
-        overlay="Visibility"
-        overlayClassName="c-rc-tooltip -default"
-        overlayStyle={{
-          color: '#fff'
-        }}
-        placement="top"
-        trigger={['hover', 'click']}
-        mouseLeaveDelay={0}
-        destroyTooltipOnHide
-      >
-        <button
-          type="button"
-          className="toggle"
-          onClick={() => this.props.onChangeVisibility(activeLayer, !visible)}
-          aria-label="Toggle the visibility"
-        >
-          <Icon name={visible ? 'icon-hide' : 'icon-show'} />
-        </button>
-      </Tooltip>
-    );
-  }
-}
-
-export default LegendItemButtonVisibility;
