+++ src/components/legend/legend-list/legend-item/legend-item-buttons/styles.scss
@@ -1,0 +1,16 @@
+@import 'settings';
+
+.c-legend-item-buttons {
+  margin-left: 20px; // Margin with the title
+  flex-shrink: 0;
+  display: flex;
+  align-items: center;
+
+  > * {
+    margin-right: 12px;
+
+    &:last-child {
+      margin-right: 0;
+    }
+  }
+}
