+++ src/components/legend/legend-list/legend-item/legend-item-type/styles.scss
@@ -1,0 +1,72 @@
+@import 'settings';
+
+.c-legend-type {
+  .type {
+    margin: 8px 0 0;
+
+    .type-list {
+      display: flex;
+      flex-flow: row wrap;
+
+      &:last-child { margin-top: 6px; }
+
+      .type-list-item {
+        .color {
+          display: block;
+          height: 8px;
+        }
+
+        .value, .name {
+          display: block;
+          margin-left: 8px;
+          line-height: 1.25;
+          color: $color-dark-1;
+          font-size: 11px;
+          font-weight: 400;
+
+          &.-dark { color: $color-dark-1; }
+        }
+      }
+    }
+
+    // GRADIENT
+    &.-gradient {
+      .type-list {
+        justify-content: space-between;
+      }
+    }
+
+    // BASIC
+    &.-basic {
+      margin: -5px 0 0;
+
+      .type-list-item {
+        display: flex;
+        justify-content: flex-start;
+        align-items: flex-start;
+        margin: 5px 15px 0 0;
+
+        .color {
+          width: 12px;
+          height: 9px;
+          margin-top: 2px;
+        }
+      }
+    }
+
+    // CHOROPLETH
+    &.-choropleth {
+      .type-list-item {
+        > .value {
+          display: inline-block;
+          width: 100%;
+          margin: 0;
+          padding: 2px;
+          text-align: center;
+          text-overflow: ellipsis;
+          overflow: hidden;
+        }
+      }
+    }
+  }
+}
