+++ src/components/legend/legend-list/styles.scss
@@ -1,3 +1,5 @@
+@import 'settings';
+
 .c-legend-list {
   margin: 0;
   padding: 0;
