+++ styleguide.template.html
@@ -8,6 +8,10 @@
 
     <title><%= htmlWebpackPlugin.options.title %></title>
     <style>
+      *, *:after, *:before {
+        box-sizing: border-box;
+      }
+
       body {
         background: white;
